Prepare for Stylelint 16.0.0
- Change the `loadLint` option's default value.
- Update the `stylelint` dev dependency (`npm i stylelint@next --save-dev`).
- Migrate test files to ESM, although files executed by Jest are still CJS.
  (The current Jest's support for ESM is still experimental)
- Update TypeScript config for newer Node.js versions.
