+++ .github/workflows/nodejs.yml
@@ -14,5 +14,3 @@ jobs:
 
   test:
     uses: stylelint/.github/.github/workflows/test.yml@main
-    with:
-      node-version: '["14", "16", "18", "20"]'
