+++ __tests__/index.test.mjs
@@ -1,6 +1,4 @@
-'use strict';
-
-const preset = require('../jest-preset.js');
+import preset from '../jest-preset.js';
 
 it('should contain expected keys', () => {
 	expect(Object.keys(preset)).toMatchSnapshot();
