+++ tsconfig.json
@@ -1,8 +1,9 @@
 {
 	"compilerOptions": {
-		"target": "ES2020",
-		"lib": ["ES2020"],
-		"module": "commonjs",
+		"target": "ES2022",
+		"lib": ["ES2022"],
+		"module": "CommonJS",
+		"moduleResolution": "Node",
 		"allowJs": true,
 		"checkJs": true,
 		"noEmit": true,
